+++ CHANGELOG.md
@@ -17,11 +17,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - `discard()` method to `AutoByteArray` and `AutoPrimitiveArray`. (#275)
-- deprecation notice for get/release/commit_byte/primitive_array_{elements|critical}. (#275)
 - Added `AutoLongArray`. (#276)
 
 ### Changed
 - `AutoByte/PrimitiveArray.commit()` now returns `Result`. (#275)
+- Removed methods get/release/commit_byte/primitive_array_{elements|critical}. (#281)
+- Renamed methods get_auto_byte/long/primitive_array_{elements|critical} to
+	get_byte/long/primitive_array_{elements|critical}. (#281)
 
 ## [0.18.0] — 2020-09-23